+++ graphql-maven-plugin-logic/src/main/resources/templates/server_GraphQLDataFetchers.vm.java
@@ -37,11 +37,11 @@ public class GraphQLDataFetchers {
 
 #foreach ($dataFetchersDelegate in $dataFetchersDelegates)
 	@Resource
-	${dataFetchersDelegate.pascalCaseName} ${dataFetchersDelegate.camelCaseName};
+	protected ${dataFetchersDelegate.pascalCaseName} ${dataFetchersDelegate.camelCaseName};
 
 #end
 	@Resource
-	GraphqlUtils graphqlUtils;
+	protected GraphqlUtils graphqlUtils;
 
 #foreach ($dataFetchersDelegate in $dataFetchersDelegates)
 	////////////////////////////////////////////////////////////////////////////////////////////////
@@ -74,7 +74,7 @@ public class GraphQLDataFetchers {
 						() -> ${dataFetchersDelegate.camelCaseName}.${dataFetcher.javaName}(dataFetchingEnvironment#if($dataFetcher.graphQLOriginType), source#end#foreach($argument in $dataFetcher.field.inputParameters), ${argument.javaName}#end));
 			}
 #elseif (${dataFetcher.field.fieldTypeAST.list})
-			#if (${dataFetcher.dataFetcherDelegate.type.requestType}=="subscription")Publisher<#end ${dataFetcher.field.javaType}#if(${dataFetcher.dataFetcherDelegate.type.requestType}=="subscription")>#end ret = ${dataFetchersDelegate.camelCaseName}.${dataFetcher.javaName}(dataFetchingEnvironment#if($dataFetcher.graphQLOriginType), source#end#foreach($argument in $dataFetcher.field.inputParameters), ${argument.javaName}#end);
+			#if (${dataFetcher.dataFetcherDelegate.type.requestType}=="subscription")Publisher<#end${dataFetcher.field.javaType}#if(${dataFetcher.dataFetcherDelegate.type.requestType}=="subscription")>#end ret = ${dataFetchersDelegate.camelCaseName}.${dataFetcher.javaName}(dataFetchingEnvironment#if($dataFetcher.graphQLOriginType), source#end#foreach($argument in $dataFetcher.field.inputParameters), ${argument.javaName}#end);
 #if (${dataFetcher.dataFetcherDelegate.type.requestType}!="subscription")
 			logger.debug("${dataFetcher.name}: {} found rows", (ret==null) ? 0 : ret.size());
 #end