+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-Forum-server/src/main/java/org/forum/server/specific_code/CustomGraphQLDataFetchers.java
@@ -1,0 +1,53 @@
+/**
+ * 
+ */
+package org.forum.server.specific_code;
+
+import java.util.NoSuchElementException;
+
+import org.forum.server.graphql.Board;
+import org.springframework.context.annotation.Primary;
+import org.springframework.stereotype.Component;
+
+import graphql.schema.DataFetcher;
+
+/**
+ * A Data Fetcher customization, that override some of the generated data fetchers. This bean is marked as
+ * {@link Primary}, and will replace the generated bean. As this class inherits from the
+ * {@link org.forum.server.graphql.GraphQLDataFetchers}, all non override data fetchers will work as expected by the
+ * generated. Only tyhe overriden method will change the generated behavior, for these data fetchers.
+ * 
+ * @author etienne-sf
+ */
+@Component("overridenGraphQLDataFetchers")
+@Primary
+public class CustomGraphQLDataFetchers extends org.forum.server.graphql.GraphQLDataFetchers {
+
+	@Override
+	public DataFetcher<Board> dataFetchersDelegateMutationTypeCreateBoard() {
+		return dataFetchingEnvironment -> {
+			String name = (String) graphqlUtils.getArgument(dataFetchingEnvironment.getArgument("name"),
+					"${argument.type.graphQLTypeSimpleName}", "java.lang.Long", String.class);
+			Boolean publiclyAvailable = (Boolean) graphqlUtils.getArgument(
+					dataFetchingEnvironment.getArgument("publiclyAvailable"), "${argument.type.graphQLTypeSimpleName}",
+					"java.lang.Long", Boolean.class);
+
+			Board ret = null;
+			try {
+				// HERE IS WHAT's CHANGED IN THIS OVERRIDEN VERSION:
+				// We add " (Overriden DataFetcher)" to the name
+				ret = dataFetchersDelegateMutationType.createBoard(dataFetchingEnvironment,
+						name + " (Overriden DataFetcher)", publiclyAvailable);
+			} catch (NoSuchElementException e) {
+				// There was no items in the Optional
+			}
+
+			if (ret != null)
+				logger.debug("createBoard (Overriden DataFetcher): 1 result found");
+			else
+				logger.debug("createBoard (Overriden DataFetcher): no result found");
+
+			return ret;
+		};
+	}
+}